requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
tensorflow
Pillow==10.3.0
numpy
